making some progress on balance stability
