+++ src/components/Profile.tsx
@@ -1,31 +1,20 @@
-import { useState, useEffect, useContext } from 'react'
+import { useState, useEffect, useContext, useCallback } from 'react'
 import AmountDisplay from './AmountDisplay'
 // import confederacyHost from '../utils/confederacyHost'
-import { makeStyles } from '@mui/styles'
-import { Typography } from '@mui/material'
+import { Stack, Typography } from '@mui/material'
 import { WalletContext } from '../WalletContext'
-
-const useStyles = makeStyles((theme: any) => ({
-  content_wrap: {
-    marginTop: '3em',
-    zIndex: 3,
-    display: 'grid',
-    placeItems: 'center',
-    paddingBottom: theme.spacing(2)
-  },
-  manage_link: {
-    textDecoration: 'underline'
-  }
-}), { name: 'Profile' })
+import AppLogo from './AppLogo'
 
 const Profile = () => {
   const { managers, adminOriginator } = useContext(WalletContext)
   const [accountBalance, setAccountBalance] = useState<any>(null)
   const [balanceLoading, setBalanceLoading] = useState(true)
-  const classes = useStyles()
 
-  const refreshBalance = async () => {
+  const refreshBalance = useCallback(async () => {
     try {
+      if (!managers?.permissionsManager) {
+        return
+      }
       setBalanceLoading(true)
       const limit = 10000
       let offset = 0
@@ -49,38 +38,31 @@ const Profile = () => {
     } catch (e) {
       setBalanceLoading(false)
     }
-  }
+  }, [managers, adminOriginator])
 
   useEffect(() => {
-    (async () => {
-      if (typeof adminOriginator === 'string') {
-        try {
-          refreshBalance()
-        } catch (e) { }
-      }
-    })()
-  }, [adminOriginator])
+    refreshBalance()
+  }, [refreshBalance])
 
-  return (
-    <>
-      <div className={classes.content_wrap}>
-        <Typography variant='h5' color='textSecondary'>
-          Your Balance
-        </Typography>
-        <Typography
-          onClick={() => refreshBalance()}
-          color='textPrimary'
-          variant='h2'
-          style={{ cursor: 'pointer' }}
-        >
-          {balanceLoading
-            ? '---'
-            : <AmountDisplay abbreviate>{accountBalance}</AmountDisplay>}
-        </Typography>
-        {/* <a href='#' className={classes.manage_link}>manage</a> */}
-      </div>
-    </>
-  )
+  if (!managers?.permissionsManager) {
+    return <AppLogo size={200} rotate />
+  }
+
+  return (<Stack>
+    <Typography variant='h5' color='textSecondary'>
+      Your Balance
+    </Typography>
+    <Typography
+      onClick={() => refreshBalance()}
+      color='textPrimary'
+      variant='h2'
+      style={{ cursor: 'pointer' }}
+    >
+      {balanceLoading
+        ? '---'
+        : <AmountDisplay abbreviate>{accountBalance}</AmountDisplay>}
+    </Typography>
+  </Stack>)
 }
 
 export default Profile
